@@ -1,43 +1,48 @@
 # coding: utf-8
 source 'http://ruby.taobao.org'
 
-gem "rails", "3.2.13"
+gem "rails", "4.0.2"
+gem 'turbolinks', '~> 1.2.0'
+gem 'jquery-turbolinks', '2.0.0'
+gem 'actionpack-action_caching', '1.0.0'
+gem 'sass-rails', "~> 4.0.0"
+gem 'coffee-rails', "~> 4.0.0"
+gem 'uglifier', '>= 1.3.0'
+
 gem "rails-i18n","0.1.8"
-gem "jquery-rails", "2.0.1"
-gem "rails_autolink", ">= 1.0.9"
-gem "jquery-atwho-rails", "0.1.6"
-gem "cache_digests"
+gem "jquery-rails", "3.0.4"
+gem "rails_autolink", ">= 1.1.0"
 gem "md_emoji"
 gem 'exception_notification'
 
 # 上传组件
 gem 'carrierwave', '0.6.2'
 gem 'carrierwave-upyun', '0.1.5'
-gem 'mini_magick','3.3', :require => false
+gem 'mini_magick','3.3', require: false
 
 # Mongoid 辅助插件
-gem "mongoid", "3.1.1"
-gem 'mongoid_auto_increment_id', "0.5.1"
+gem "mongoid", github: 'mongoid/mongoid', ref: '11e45e5a30a45458b83db99ab6c9d9ccc337e66f'
+gem 'mongoid_auto_increment_id', "0.6.2"
 gem 'mongoid_rails_migrations', '1.0.0'
 
 # 用户系统
-gem 'devise', '2.2.3'
-gem 'devise-encryptable', '0.1.1'
+gem 'devise', '3.0.1'
+gem 'devise-encryptable', '0.1.2'
 
 # 分页
-gem 'will_paginate', '3.0.2'
+gem 'will_paginate', '3.0.4'
 
 # Bootstrap
-gem 'anjlab-bootstrap-rails', '2.0.3.2', :require => 'bootstrap-rails'
+gem 'anjlab-bootstrap-rails', '2.0.3.2', require: 'bootstrap-rails'
 gem 'bootstrap-will_paginate', '0.0.3'
-gem 'bootstrap_helper', "1.4.1"
+gem 'bootstrap_helper', github: 'huacnlee/bootstrap-helper'
 
 # 三方平台 OAuth 验证登陆
 gem "omniauth", "~> 1.0.1"
-gem "omniauth-github", "~> 1.0.0"
+gem "omniauth-github", "~> 1.1.0"
 
 # permission
-gem "cancan", "~> 1.6.7"
+gem "cancan", "~> 1.6.10"
 
 # Redis 命名空间
 gem 'redis-namespace','~> 1.2.1'
