@@ -16,7 +16,7 @@ class User
   # Email 的 md5 值，用于 Gravatar 头像
   field :email_md5
   # Email 是否公开
-  field :email_public, :type => Boolean
+  field :email_public, :type => Mongoid::Boolean
   field :encrypted_password, :type => String, :default => ""
 
   validates_presence_of :email
@@ -45,9 +45,9 @@ class User
   field :github
   field :twitter
   # 是否信任用户
-  field :verified, :type => Boolean, :default => false
+  field :verified, :type => Mongoid::Boolean, :default => false
   field :state, :type => Integer, :default => 1
-  field :guest, :type => Boolean, :default => false
+  field :guest, :type => Mongoid::Boolean, :default => false
   field :tagline
   field :topics_count, :type => Integer, :default => 0
   field :replies_count, :type => Integer, :default => 0
@@ -76,12 +76,12 @@ class User
         :user_id => id,
         :_id.in  => unread_ids,
         :read    => false
-      }).update_all(:read => true, :updated_at => Time.now)
+      }).update_all(read: true, updated_at: Time.now)
     end
   end
 
   attr_accessor :password_confirmation
-  attr_accessible :name, :email_public, :location, :company, :bio, :website, :github, :twitter, :tagline, :avatar, :password, :password_confirmation
+  ACCESSABLE_ATTRS = [:name, :email_public, :location, :company, :bio, :website, :github, :twitter, :tagline, :avatar, :by, :current_password, :password, :password_confirmation]
 
   validates :login, :format => {:with => /\A\w+\z/, :message => '只允许数字、大小写字母和下划线'}, :length => {:in => 3..20}, :presence => true, :uniqueness => {:case_sensitive => false}
 
@@ -95,7 +95,7 @@ class User
     self.email_md5 = Digest::MD5.hexdigest(val || "")
     self[:email] = val
   end
-  
+
   def temp_access_token
     Rails.cache.fetch("user-#{self.id}-temp_access_token-#{Time.now.strftime("%Y%m%d")}") do
       SecureRandom.hex
@@ -136,16 +136,24 @@ class User
   def wiki_editor?
     self.admin? or self.verified == true
   end
-  
+
   # 回帖大于 150 的才有酷站的发布权限
   def site_editor?
     self.admin? or self.replies_count >= 100
   end
-  
+
   # 是否能发帖
   def newbie?
     return false if self.verified == true
     self.created_at > 1.week.ago
+  end
+
+  def blocked?
+    return self.state == STATE[:blocked]
+  end
+
+  def deleted?
+    return self.state == STATE[:deleted]
   end
 
   def has_role?(role)
@@ -153,15 +161,15 @@ class User
       when :admin then admin?
       when :wiki_editor then wiki_editor?
       when :site_editor then site_editor?
-      when :member then true
+      when :member then self.state == STATE[:normal]
       else false
     end
   end
 
-  before_create :default_value_for_create
-  def default_value_for_create
-    self.state = STATE[:normal]
-  end
+  # before_create :default_value_for_create
+  # def default_value_for_create
+  #   self.state = STATE[:normal]
+  # end
 
   # 注册邮件提醒
   after_create :send_welcome_mail
@@ -175,9 +183,9 @@ class User
     if self.location_changed?
       if not self.location.blank?
         old_location = Location.find_by_name(self.location_was)
-        old_location.inc(:users_count, -1) if not old_location.blank?
+        old_location.inc(users_count: -1) if not old_location.blank?
         location = Location.find_or_create_by_name(self.location)
-        location.inc(:users_count, 1)
+        location.inc(users_count: 1)
         self.location_id = (location.blank? ? nil : location.id)
       else
         self.location_id = nil
@@ -213,7 +221,7 @@ class User
 
   def bind_service(response)
     provider = response["provider"]
-    uid = response["uid"]
+    uid = response["uid"].to_s
     authorizations.create(:provider => provider , :uid => uid )
   end
 
@@ -228,11 +236,11 @@ class User
   def read_topic(topic)
     return if topic.blank?
     return if self.topic_read?(topic)
-    
+
     self.notifications.unread.any_of({:mentionable_type => 'Topic', :mentionable_id => topic.id},
                                      {:mentionable_type => 'Reply', :mentionable_id.in => topic.reply_ids},
-                                     {:reply_id.in => topic.reply_ids}).update_all(:read => true)
-    
+                                     {:reply_id.in => topic.reply_ids}).update_all(read: true)
+
     # 处理 last_reply_id 是空的情况
     last_reply_id = topic.last_reply_id || -1
     Rails.cache.write("user:#{self.id}:topic_read:#{topic.id}", last_reply_id)
@@ -242,8 +250,8 @@ class User
   def like(likeable)
     return false if likeable.blank?
     return false if likeable.liked_by_user?(self)
-    likeable.push(:liked_user_ids, self.id)
-    likeable.inc(:likes_count, 1)
+    likeable.push(liked_user_ids: self.id)
+    likeable.inc(likes_count: 1)
     likeable.touch
   end
 
@@ -251,8 +259,8 @@ class User
   def unlike(likeable)
     return false if likeable.blank?
     return false if not likeable.liked_by_user?(self)
-    likeable.pull(:liked_user_ids, self.id)
-    likeable.inc(:likes_count, -1)
+    likeable.pull(liked_user_ids: self.id)
+    likeable.inc(likes_count: -1)
     likeable.touch
   end
 
@@ -261,7 +269,7 @@ class User
     return false if topic_id.blank?
     topic_id = topic_id.to_i
     return false if self.favorite_topic_ids.include?(topic_id)
-    self.push(:favorite_topic_ids, topic_id)
+    self.push(favorite_topic_ids: topic_id)
     true
   end
 
@@ -269,7 +277,7 @@ class User
   def unfavorite_topic(topic_id)
     return false if topic_id.blank?
     topic_id = topic_id.to_i
-    self.pull(:favorite_topic_ids, topic_id)
+    self.pull(favorite_topic_ids: topic_id)
     true
   end
 
@@ -325,7 +333,7 @@ class User
     random_key = "#{SecureRandom.hex(10)}:#{self.id}"
     self.update_attribute(:private_token, random_key)
   end
-  
+
   def ensure_private_token!
     self.update_private_token if self.private_token.blank?
   end
