@@ -1,5 +1,5 @@
 # coding: utf-8
-class AccountController < Devise::RegistrationsController
+class AccountController < Devise::RegistrationsController  
   def edit
     @user = current_user
     # 首次生成用户 Token
@@ -7,30 +7,12 @@ class AccountController < Devise::RegistrationsController
   end
 
   def update
-    self.resource = resource_class.to_adapter.get!(send(:"current_#{resource_name}").to_key)
-
-    # Can update Email when email was not has being exist.
-    resource.email = resource_params[:email] if self.resource.email.blank?
-
-    # code from Devise
-    if resource.update_with_password(resource_params)
-      if is_navigational_format?
-        if resource.respond_to?(:pending_reconfirmation?) && resource.pending_reconfirmation?
-          flash_key = :update_needs_confirmation
-        end
-        set_flash_message :notice, flash_key || :updated
-      end
-      sign_in resource_name, resource, :bypass => true
-      respond_with resource, :location => after_update_path_for(resource)
-    else
-      clean_up_passwords resource
-      respond_with resource
-    end
+    super
   end
 
   # POST /resource
   def create
-    build_resource
+    build_resource(sign_up_params)
     resource.login = params[resource_name][:login]
     resource.email = params[resource_name][:email]
     if resource.save
