@@ -6,7 +6,7 @@ class Cpanel::RepliesController < Cpanel::ApplicationController
   end
 
   def show
-    @reply = Reply.find(params[:id])
+    @reply = Reply.unscoped.find(params[:id])
 
     if @reply.topic.blank?
       redirect_to cpanel_replies_path, :alert => "帖子已经不存在"
@@ -18,12 +18,12 @@ class Cpanel::RepliesController < Cpanel::ApplicationController
   end
 
   def edit
-    @reply = Reply.find(params[:id])
+    @reply = Reply.unscoped.find(params[:id])
   end
 
 
   def create
-    @reply = Reply.new(params[:reply])
+    @reply = Reply.new(params[:reply].permit!)
 
     if @reply.save
       redirect_to(cpanel_replies_path, :notice => 'Reply was successfully created.')
@@ -33,9 +33,9 @@ class Cpanel::RepliesController < Cpanel::ApplicationController
   end
 
   def update
-    @reply = Reply.find(params[:id])
+    @reply = Reply.unscoped.find(params[:id])
 
-    if @reply.update_attributes(params[:reply])
+    if @reply.update_attributes(params[:reply].permit!)
        redirect_to(cpanel_replies_path, :notice => 'Reply was successfully updated.')
     else
       render :action => "edit"
@@ -43,9 +43,7 @@ class Cpanel::RepliesController < Cpanel::ApplicationController
   end
 
   def destroy
-    @reply = Reply.find(params[:id])
+    @reply = Reply.unscoped.find(params[:id])
     @reply.destroy
-
-    redirect_to(cpanel_replies_path)
   end
 end
