@@ -12,7 +12,7 @@ class Cpanel::SiteConfigsController < Cpanel::ApplicationController
   def update
     @site_config = SiteConfig.find(params[:id])
 
-    if @site_config.update_attributes(params[:site_config])
+    if @site_config.update_attributes(params[:site_config].permit!)
       redirect_to edit_cpanel_site_config_path(params[:id]), notice: "保存成功."
     else
       render action: "edit"
