@@ -20,7 +20,7 @@ class Cpanel::UsersController < Cpanel::ApplicationController
   end
 
   def create
-    @user = User.new(params[:user])
+    @user = User.new(params[:user].permit!)
     @user.email = params[:user][:email]
     @user.login = params[:user][:login]
     @user.state = params[:user][:state]
@@ -40,8 +40,8 @@ class Cpanel::UsersController < Cpanel::ApplicationController
     @user.state = params[:user][:state]
     @user.verified = params[:user][:verified]
 
-    if @user.update_attributes(params[:user])
-      redirect_to(cpanel_users_path, :notice => 'User was successfully updated.')
+    if @user.update_attributes(params[:user].permit!)
+      redirect_to(edit_cpanel_user_path(@user.id), :notice => 'User was successfully updated.')
     else
       render :action => "edit"
     end
