@@ -1,8 +1,7 @@
 # coding: utf-8
 class HomeController < ApplicationController
-  caches_action :index, :expires_in => 1.hours, :layout => false
-
   def index
+    @excellent_topics = Topic.excellent.recent.fields_for_list.includes(:user).limit(20)
     drop_breadcrumb("首页", root_path)
   end
 
