@@ -1,6 +1,8 @@
 # coding: utf-8
 class SearchController < ApplicationController
   def index
-    redirect_to "https://www.google.com.hk/#hl=zh-CN&q=site:ruby-china.org+#{params[:q]}"
+    keywords = params[:q] || ""
+    keywords.gsub!("#", "%23")
+    redirect_to "https://www.google.com.hk/#hl=zh-CN&q=site:ruby-china.org+#{keywords}"
   end
 end
