@@ -15,7 +15,7 @@ class SitesController < ApplicationController
   end
 
   def create
-    @site = Site.new(params[:site])
+    @site = Site.new(site_params)
     @site.user_id = current_user.id
     if @site.save
       redirect_to(sites_path, :notice => '提交成功。谢谢。')
@@ -26,4 +26,9 @@ class SitesController < ApplicationController
     end
   end
 
+  private
+  
+  def site_params
+    params.require(:site).permit(:name,:desc,:url,:favorite,:site_node_id)
+  end
 end
