@@ -15,7 +15,11 @@ module RubyChina
     end
 
     class DetailUser < Grape::Entity
-      expose :id, :name, :login, :email, :location, :company, :twitter, :website, :bio, :tagline, :github_url
+      expose :id, :name, :login, :location, :company, :twitter, :website, :bio, :tagline, :github_url
+      expose :email do |model, opts|
+        model.email_public ? model.email : ''
+      end
+
       # deprecated: gravatar_hash, use avatar_url for user avatar
       expose(:gravatar_hash) { |model, opts| Digest::MD5.hexdigest(model.email || "") }
       expose(:avatar_url) do |model, opts|
