@@ -27,7 +27,12 @@ class Ability
       end
 
       # Reply
-      can :create, Reply
+      # 新手用户晚上禁止回帖，防 spam，可在面板设置是否打开
+      if !(user.newbie? &&
+           (SiteConfig.reject_newbie_reply_in_the_evening == 'true') &&
+           (Time.zone.now.hour < 9 || Time.zone.now.hour > 22))
+        can :create, Reply
+      end
       can :update, Reply do |reply|
         reply.user_id == user.id
       end
@@ -61,7 +66,6 @@ class Ability
         end
       end
 
-
       # Photo
       can :tiny_new, Photo
       can :create, Photo
@@ -92,7 +96,6 @@ class Ability
       cannot :manage, :all
       basic_read_only
     end
-
 
   end
 
