@@ -6,7 +6,7 @@ module Mongoid
 
     included do
       scope :recent, desc(:_id)
-      scope :exclude_ids, Proc.new { |ids| where(:_id.nin => ids.map(&:to_i)) }
+      scope :exclude_ids, lambda { |ids| where(:_id.nin => ids.map(&:to_i)) }
       scope :by_week, where(:created_at.gte => 7.days.ago.utc)
       
       delegate :url_helpers, to: 'Rails.application.routes' 
