@@ -25,14 +25,14 @@ module Mongoid
         set_callback(:create, :after) do |document|
           relation = document.send(metadata[:name])
           if relation
-            relation.inc(counter_name.to_sym, 1) if relation.class.fields.keys.include?(counter_name)
+            relation.inc(counter_name.to_sym => 1) if relation.class.fields.keys.include?(counter_name)
           end
         end
 
         set_callback(:destroy, :after) do |document|
           relation = document.send(metadata[:name])
           if relation && relation.class.fields.keys.include?(counter_name)
-            relation.inc(counter_name.to_sym, -1)
+            relation.inc(counter_name.to_sym => -1)
           end
         end
 
