@@ -1,6 +1,6 @@
 class Notification::Base
   include Mongoid::Document
-  include Mongoid::Timestamps::Created
+  include Mongoid::Timestamps
   include Mongoid::BaseModel
 
   store_in :collection => 'notifications'
