@@ -1,47 +1,28 @@
 # coding: utf-8
 require "bundler/capistrano"
 require "sidekiq/capistrano"
-
 require "rvm/capistrano"
-default_run_options[:pty] = true
-set :rvm_ruby_string, 'ruby-2.0.0-p0'
-set :rvm_type, :user
+require 'puma/capistrano'
 
+default_run_options[:pty] = true
+
+set :rvm_ruby_string, 'ruby-2.1.0'
+set :rvm_type, :user
 set :application, "ruby-china"
 set :repository,  "git://github.com/ruby-china/ruby-china.git"
 set :branch, "master"
 set :scm, :git
 set :user, "ruby"
-if ENV["DEPLOY"] == "pre"
-  set :deploy_to, "/data/www/#{application}-pre"
-else
-  set :deploy_to, "/data/www/#{application}"
-end
+set :deploy_to, "/data/www/#{application}"
 set :runner, "ruby"
 # set :deploy_via, :remote_cache
-# set :git_shallow_clone, 1
+set :git_shallow_clone, 1
+set :puma_role, :app
+set :puma_config_file, "config/puma.rb"
 
-role :web, "s2.ruby-china.org"                          # Your HTTP server, Apache/etc
-role :app, "s2.ruby-china.org"                          # This may be the same as your `Web` server
-role :db,  "s2.ruby-china.org", :primary => true # This is where Rails migrations will run
-
-# unicorn.rb 路径
-set :unicorn_path, "#{deploy_to}/current/config/unicorn.rb"
-
-namespace :deploy do
-  task :start, :roles => :app do
-    run "cd #{deploy_to}/current/; RAILS_ENV=production bundle exec unicorn_rails -c #{unicorn_path} -D"
-  end
-
-  task :stop, :roles => :app do
-    run "kill -QUIT `cat #{deploy_to}/current/tmp/pids/unicorn.pid`"
-  end
-
-  desc "Restart Application"
-  task :restart, :roles => :app do
-    run "kill -USR2 `cat #{deploy_to}/current/tmp/pids/unicorn.pid`"
-  end
-end
+role :web, "ruby-china.org"                          # Your HTTP server, Apache/etc
+role :app, "ruby-china.org"                          # This may be the same as your `Web` server
+role :db,  "ruby-china.org", :primary => true # This is where Rails migrations will run
 
 namespace :faye do
   desc "Start Faye"
@@ -60,7 +41,6 @@ namespace :faye do
   end
 end
 
-
 task :init_shared_path, :roles => :web do
   run "mkdir -p #{deploy_to}/shared/log"
   run "mkdir -p #{deploy_to}/shared/pids"
@@ -68,10 +48,11 @@ task :init_shared_path, :roles => :web do
 end
 
 task :link_shared_files, :roles => :web do
+  run "ln -sf #{shared_path}/assets #{deploy_to}/current/public/assets"
   run "ln -sf #{deploy_to}/shared/config/*.yml #{deploy_to}/current/config/"
-  run "ln -sf #{deploy_to}/shared/config/unicorn.rb #{deploy_to}/current/config/"
   run "ln -sf #{deploy_to}/shared/config/initializers/secret_token.rb #{deploy_to}/current/config/initializers"
   run "ln -sf #{deploy_to}/shared/config/faye_thin.yml #{deploy_to}/current/faye_server/thin.yml"
+  run "ln -sf #{shared_path}/pids #{deploy_to}/current/tmp/"
 end
 
 task :mongoid_create_indexes, :roles => :web do
@@ -80,9 +61,6 @@ end
 
 task :compile_assets, :roles => :web do
   run "cd #{deploy_to}/current/; RAILS_ENV=production bundle exec rake assets:precompile"
-end
-
-task :sync_assets_to_cdn, :roles => :web do
   run "cd #{deploy_to}/current/; RAILS_ENV=production bundle exec rake assets:cdn"
 end
 
@@ -90,4 +68,4 @@ task :mongoid_migrate_database, :roles => :web do
   run "cd #{deploy_to}/current/; RAILS_ENV=production bundle exec rake db:migrate"
 end
 
-after "deploy:finalize_update","deploy:symlink", :init_shared_path, :link_shared_files#, :compile_assets, :sync_assets_to_cdn, :mongoid_migrate_database
+after "deploy:finalize_update","deploy:symlink", :init_shared_path, :link_shared_files, :mongoid_migrate_database #, :compile_assets
