@@ -2,8 +2,7 @@
 # Be sure to restart your server when you modify this file.
 
 RubyChina::Application.config.session_store :cookie_store,
-                            :key => '_homeland_session',
-                            :expire_after => 86400*90
+                            :key => '_homeland_session'
 
 # Use the database for sessions instead of the cookie-based default,
 # which shouldn't be used to store highly confidential information
