@@ -1,5 +1,5 @@
 # coding: utf-8
-require File.expand_path('../../../lib/markdown', __FILE__)
+require 'spec_helper'
 
 describe 'markdown' do
   let(:upload_url) { '' }
@@ -114,6 +114,20 @@ describe 'markdown' do
         specify { doc.css('a').should be_empty }
         specify { doc.css('pre').inner_html.should == "@user\n" }
       end
+
+      context '@var in coffeescript' do
+        let(:raw) {
+          <<-MD.gsub(/^ {12}/, '')
+            ```coffeescript
+            @var
+            ```
+          MD
+        }
+
+        it 'should not leave it as placeholder' do
+          doc.to_html.should include('var')
+        end
+      end
     end
 
     # }}}
@@ -211,5 +225,56 @@ describe 'markdown' do
     end
 
     # }}}
+    
+    describe 'The code' do
+      context '``` use with code' do
+        let(:raw) {
+          %(```
+          class Foo; end
+          ```)
+        }
+        
+        specify { doc.css('pre').attr("class").value.should == "highlight plaintext" }
+      end
+      
+      context '```ruby use with code' do
+        let(:raw) {
+          %(```ruby
+          class Foo; end
+          ```)
+        }
+        
+        specify { doc.css('pre').attr("class").value.should == "highlight ruby" }
+      end
+      
+      context 'indent in raw with \t' do
+        let(:raw) { "\t\tclass Foo; end" }
+        
+        specify { doc.css('pre').should be_empty }
+      end
+      
+      context 'indent in raw with space' do
+        let(:raw) { "    class Foo; end" }
+        
+        specify { doc.css('pre').should be_empty }
+      end
+    end
+    
+    describe 'Escape HTML tags' do
+      context '<xxx> or a book names' do
+        let(:raw) { "<Enterprise Integration Patterns> book" }
+        its(:inner_html) { should == "<p>&lt;Enterprise Integration Patterns&gt; book</p>" }
+      end
+      
+      context '<img> tag' do
+        let(:raw) { "<img src='aaa.jpg' /> aaa" }
+        its(:inner_html) { should == "<p>&lt;img src='aaa.jpg' /&gt; aaa</p>" }
+      end
+      
+      context '<b> tag' do
+        let(:raw) { "<b>aaa</b>" }
+        its(:inner_html) { should == "<p>&lt;b&gt;aaa&lt;/b&gt;</p>" }
+      end
+    end
   end
 end
