@@ -94,21 +94,21 @@ describe User do
       user.should_not be_wiki_editor
     end
   end
-  
+
   describe "newbie?" do
     it "should true when user created_at less than a week" do
       user.verified = false
       user.created_at = 6.days.ago
       user.newbie?.should be_true
     end
-    
+
     it "should false when more than a week and have 10+ replies" do
       user.verified = false
       user.created_at = 10.days.ago
       user.replies_count = 10
       user.newbie?.should be_false
     end
-    
+
     it "should false when user is verified" do
       user.verified = true
       user.newbie?.should be_false
@@ -121,6 +121,16 @@ describe User do
     context "when is a new user" do
       let(:user) { Factory :user }
       it { should have_role(:member) }
+    end
+
+    context "when is a blocked user" do
+      let(:user) { Factory :blocked_user }
+      it { should_not have_role(:member) }
+    end
+
+    context "when is a deleted user" do
+      let(:user) { Factory :blocked_user }
+      it { should_not have_role(:member) }
     end
 
     context "when is admin" do
